+++ sesh/sesh.go
@@ -1,11 +1,17 @@
 package sesh
 
 import (
+	"fmt"
+	"slices"
+
+	"github.com/charmbracelet/log"
+	com "github.com/garrettkrohn/treekanga/common"
 	"github.com/garrettkrohn/treekanga/shell"
+	"github.com/garrettkrohn/treekanga/utility"
 )
 
 type Sesh interface {
-	SeshConnect(seshName string) error
+	SeshConnect(c *com.AddConfig)
 }
 
 type RealSesh struct {
@@ -16,7 +22,22 @@ func NewSesh(shell shell.Shell) Sesh {
 	return &RealSesh{shell}
 }
 
-func (r *RealSesh) SeshConnect(seshName string) error {
-	_, err := r.shell.Cmd("sesh", "connect", seshName)
-	return err
+func (r *RealSesh) SeshConnect(c *com.AddConfig) {
+
+	log.Debug("connecting to: %s", c.Flags.Connect)
+
+	shortestZoxide := slices.Min(c.ZoxideConfig.FoldersToAdd)
+	subFolderIsValid := slices.Contains(c.ZoxideConfig.FoldersToAdd, *c.Flags.Connect)
+	if subFolderIsValid {
+		zoxidePath := shortestZoxide + "/" + *c.Flags.Connect
+		log.Info(fmt.Sprintf("Sesh connect to %s", zoxidePath))
+		_, err := r.shell.Cmd("sesh", "connect", zoxidePath)
+		utility.CheckError(err)
+		// deps.Sesh.SeshConnect(zoxidePath)
+	} else {
+		log.Info(fmt.Sprintf("Sesh connect to %s", shortestZoxide))
+		_, err := r.shell.Cmd("sesh", "connect", shortestZoxide)
+		utility.CheckError(err)
+		// deps.Sesh.SeshConnect(shortestZoxide)
+	}
 }
