+++ zoxide/zoxide.go
@@ -1,12 +1,18 @@
 package zoxide
 
 import (
+	"strings"
+
+	"github.com/garrettkrohn/treekanga/common"
+	"github.com/garrettkrohn/treekanga/directoryReader"
 	"github.com/garrettkrohn/treekanga/shell"
+	util "github.com/garrettkrohn/treekanga/utility"
 )
 
 type Zoxide interface {
 	AddPath(path string) error
 	RemovePath(path string) error
+	AddZoxideEntries(c *common.ZoxideConfig)
 }
 
 type RealZoxide struct {
@@ -25,4 +31,52 @@ func (r *RealZoxide) AddPath(path string) error {
 func (r *RealZoxide) RemovePath(path string) error {
 	_, err := r.shell.Cmd("zoxide", "remove", path)
 	return err
+}
+
+func (r *RealZoxide) AddZoxideEntries(c *common.ZoxideConfig) {
+	baseName := c.ParentDir + "/" + c.NewBranchName
+
+	var foldersToAdd []string
+	foldersToAdd = append(foldersToAdd, baseName)
+
+	foldersToAdd = addConfigFolders(foldersToAdd, c.FoldersToAdd, baseName, c.DirectoryReader)
+
+	for _, folder := range foldersToAdd {
+		err := r.AddPath(folder)
+		util.CheckError(err)
+	}
+}
+
+func addConfigFolders(foldersToAdd []string, foldersToAddFromConfig []string, baseName string, directoryReader directoryReader.DirectoryReader) []string {
+	for _, folder := range foldersToAddFromConfig {
+		if !isLastCharWildcard(folder) {
+			newFolderFromConfig := baseName + "/" + folder
+			foldersToAdd = append(foldersToAdd, newFolderFromConfig)
+		} else {
+			pathUpTillWildcard := getPathUntilLastSlash(folder)
+			baseFolderToSearch := baseName + "/" + pathUpTillWildcard
+			configFolders, err := directoryReader.GetFoldersInDirectory(baseFolderToSearch)
+
+			for _, configFolder := range configFolders {
+				newConfigFolder := baseFolderToSearch + "/" + configFolder
+				foldersToAdd = append(foldersToAdd, newConfigFolder)
+			}
+			util.CheckError(err)
+		}
+	}
+	return foldersToAdd
+}
+
+func getPathUntilLastSlash(input string) string {
+	parts := strings.Split(input, "/")
+	if len(parts) > 1 {
+		return strings.Join(parts[:len(parts)-1], "/")
+	}
+	return ""
+}
+
+func isLastCharWildcard(input string) bool {
+	parts := strings.Split(input, "/")
+	lastSegment := parts[len(parts)-1]
+	return strings.HasSuffix(lastSegment, "*")
 }
